maven refactor - depreicate old class names for new standerdized ones.
