+++ src/ploigos_step_runner/step_implementers/package/__init__.py
@@ -1,4 +1,5 @@
 """`StepImplementers` for the `package` step.
 """
 
+from ploigos_step_runner.step_implementers.package.maven_package import MavenPackage
 from ploigos_step_runner.step_implementers.package.maven import Maven
