+++ src/ploigos_step_runner/step_implementers/push_artifacts/__init__.py
@@ -1,4 +1,5 @@
 """`StepImplementers` for the `push-artifacts` step.
 """
 
+from ploigos_step_runner.step_implementers.push_artifacts.maven_deploy import MavenDeploy
 from ploigos_step_runner.step_implementers.push_artifacts.maven import Maven
