+++ src/ploigos_step_runner/step_implementers/uat/__init__.py
@@ -1,4 +1,7 @@
 """`StepImplementers` for the `uat` (User Acceptance Tests) step.
 """
 
-from ploigos_step_runner.step_implementers.uat.maven_selenium_cucumber import MavenSeleniumCucumber
+from ploigos_step_runner.step_implementers.uat.maven_test_selenium_cucumber import \
+    MavenTestSeleniumCucumber
+from ploigos_step_runner.step_implementers.uat.maven_selenium_cucumber import \
+    MavenSeleniumCucumber
