+++ src/ploigos_step_runner/step_implementers/unit_test/__init__.py
@@ -1,3 +1,4 @@
 """`StepImplementers` for the `unit-test` step.
 """
+from ploigos_step_runner.step_implementers.unit_test.maven_test import MavenTest
 from ploigos_step_runner.step_implementers.unit_test.maven import Maven
