+++ tests/step_implementers/package/test_maven_depricated_by_maven_package.py
@@ -1,0 +1,28 @@
+
+from unittest.mock import patch
+
+from ploigos_step_runner import WorkflowResult
+from ploigos_step_runner.step_implementers.package import Maven
+from tests.helpers.base_step_implementer_test_case import \
+    BaseStepImplementerTestCase
+
+
+@patch("ploigos_step_runner.step_implementers.package.MavenPackage.__init__")
+class TestStepImplementerMaven_UnitTestOld___init__(BaseStepImplementerTestCase):
+    def test_defaults(self, mock_super_init):
+        workflow_result = WorkflowResult()
+        parent_work_dir_path = '/fake/path'
+        config = {}
+
+        Maven(
+            workflow_result=workflow_result,
+            parent_work_dir_path=parent_work_dir_path,
+            config=config
+        )
+
+        mock_super_init.assert_called_once_with(
+            workflow_result=workflow_result,
+            parent_work_dir_path=parent_work_dir_path,
+            config=config,
+            environment=None
+        )
